+++ x/marker/keeper/genesis.go
@@ -39,6 +39,7 @@ func (k Keeper) InitGenesis(ctx sdk.Context, data *types.GenesisState) {
 					panic(err)
 				}
 			}
+
 			k.SetMarker(ctx, &data.Markers[i])
 		}
 	}