+++ CHANGELOG.md
@@ -36,6 +36,9 @@ Ref: https://keepachangelog.com/en/1.0.0/
 -->
 
 ## Unreleased
+### Improvements
+
+* Ensure marker state transition validation does not panic [#492](https://github.com/provenance-io/provenance/issues/492)
 
 ### Bug Fixes