+++ CHANGELOG.md
@@ -37,6 +37,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+## [v1.7.1](https://github.com/provenance-io/provenance/releases/tag/v1.7.1) - 2021-09-20
+
 ### Improvements
 
 * Ensure marker state transition validation does not panic [#492](https://github.com/provenance-io/provenance/issues/492)
@@ -47,6 +49,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Fixed marker with more than uint64 causes panic [#489](https://github.com/provenance-io/provenance/issues/489)
 
 ## [v1.7.0](https://github.com/provenance-io/provenance/releases/tag/v1.7.0) - 2021-09-03
+
 ### Features
 
 * Marker governance proposal are supported in cli [#367](https://github.com/provenance-io/provenance/issues/367)